Introduce new table layouts.
There are now tables for value records (split into SIMD records and scalar
records), and instruction records.  Each instruction record has a name, a
mnemonic that is used by SBCL's assembler, a list of argument value
records, a list of result value records, and some further properties.
Each instruction record is automatically translated into a VOP and a
function stub that uses this VOP.

What is still left to decide is whether packing and unpacking functions
should be represented as instructions, too, and how to encode the AREF
machinery.