Distinguish between VOP names and external names.
For each instruction, we now introduce one VOP and two functions.  One of
these functions has the same name as the VOP and simply invokes the VOP.
The other function attempts to upgrade its arguments to the correct types
before invoking the VOP.  We only export the latter.

This commit also fixes some typos in the mnemonics of some instructions.