Introduce the concept of instruction emitters.
An instruction emitter takes care of translating an instruction mnemonic
and some results and arguments to a :generator form of a VOP.  Emitters
introduce the necessary flexibility we need, e.g., for handling SIMD
comparison functions.

The only wart at the moment is that we have to use EVAL when building the
instruction table, because we want some emitters to close over things.
Better ideas are most welcome.