Change how loads and stores are defined.
Instead of having separate tables for instructions and reffers, we now
describe everything by instructions with different encodings.

The advantage of this scheme is that loads and stores can now be defined
separately and depending on the available instruction sets.  I plan to use
this later for defining non-temporal loads and stores.

I also changed the naming convention for reffers.  In the new convention,
the reffers of each SIMD type X.Y are X.Y-aref and X.Y-row-major-aref.