Introduce one package per instruction set.
Instead of always picking the newest available instructions, it is now
possible to explicitly choose one of the available instruction sets.  This
change allows users to write multiple versions of some code, and pick the
most modern version that is supported on the target platform.

Another benefit of this change is that it eliminates the need for symbols
and functions that are defined conditionally.  Instead, all functions of
all instruction sets are always defined.  Functions that are defined but
not supported signal an error when they are called.