Better SIMD packers, unpackers, and casts.  Plus some bugfixes.
SSE instructions are now always encoded correctly, even in the case where
the destination and the second operand are the same register.

We introduce functions for reinterpreting values as a particular SIMD pack.
The naming convention is X.Y! for reinterpreting the bits of an arbitrary
value as a SIMD pack of type X.Y.  If the representation of the value is
shorter than the desired pack, the missing entries are filled with zeros.
If the representation of the value is larger (e.g., when reinterpreting a
256 bit pack as a 128 bit pack), the excess elements are discarded.

Instead of using SBCL's built-in packers and unpackers, we now provide
custom versions of all MAKE-X.Y and X.Y-VALUES functions that only use the
instructions of the instruction set they are defined in.  So it is now
possible to create 256 bit packs on platforms with only AVX but not AVX2
without raising a SIGILL.

For convenience, sb-simd can now express SIMD packs with arbitrary type,
with the special SIMD value records P128 and P256.  This change was
invaluable for implementing the reinterpret casts.

Several instruction records have been added or improved.