Tweak SBCL to use VEX encoded instructions in AVX code.
This work is preliminary.  Many floating-point VOPs will still use SSE
instructions.  But some of the most important cases (like casts of reals to
floating-point numbers) are now handled correctly.

Now I need to negotiate with the SBCL maintainers how this mechanism can be
merged directly into SBCL, because a contrib is really not the best place
for it.