Define functions for element-wise conditional selection of values.
We call these functions X.Y-IF, but their semantics differs from CL:IF in
that they evaluate both branches and use a mask to choose between values of
either of them.  The rationale is that control flow doesn't mix well with
SIMD programming, and these functions are as close as one can sensibly get
to having a SIMD 'if'.

The good news is that these functions mix perfectly with the existing SIMD
conditional operators (X.Y<, X.Y=/, and the like), because those operators
return suitable bitmasks.