Avoid using untyped SIMD registers.
Currently, SBCL treats SIMD packs with unknown type as SIMD-PACK-INT.  This
gives flexibility, but undermines type dispatch and the quality of static
type checking.  There is hope that SBCL's behavior will change soon, and
this commit prepares for that change.

Also, fix a bug in the signed, scalar IF operations.