Remove vertical operations, add more horizontal operations.
This commit removes all the vdot and vsum functions.  Not because they
aren't useful, but because their implementation is beyond the scope of
sb-simd.  Ideally, those functions should be part of a tensor and linear
algebra library.

On the plus side, this commit introduces many more horizontal functions,
i.e., functions that combine the elements within a SIMD pack in an
efficient way.